Improve folder structure and organization
