+++ src/constants/data/index.ts
@@ -1,0 +1,3 @@
+export const STATUS_TABS = ['All', 'Active', 'Onboarding', 'Off-boarding', 'Dismissed'];
+
+export const EMPLOYEES_PER_PAGE = 5;
