+++ src/constants/ui/index.ts
@@ -1,0 +1,6 @@
+export const SIDEBAR_WIDTH = {
+  EXPANDED: 'w-64',
+  COLLAPSED: 'w-16',
+} as const;
+
+export const ANIMATION_DURATION = 'duration-300';
