+++ src/pages/PeoplePage.tsx
@@ -1,56 +1,19 @@
 import React, { useState, useMemo } from 'react';
+import { Plus, Download } from 'lucide-react';
+import { type SortingState, type ColumnFiltersState } from '@tanstack/react-table';
 import {
-  Plus,
-  Filter,
-  ArrowUpDown,
-  Download,
-  Search,
-  Users,
-  Eye,
-  MoreVertical,
-  Trash2,
-  Edit,
-} from 'lucide-react';
-import {
-  useReactTable,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  getFilteredRowModel,
-  flexRender,
-  type ColumnDef,
-  type SortingState,
-  type ColumnFiltersState,
-} from '@tanstack/react-table';
-import {
-  filteredEmployees as filterEmployeesUtil,
-  getUniqueDepartments,
-  getUniqueJobTitles,
-  getAvatarUrl,
-  formatDate,
+  filteredEmployees as filterEmployeesUtil
 } from '../utils';
 import type { Employee } from '../types';
 import { statusTabs } from '../constants';
 import { EmployeeModal } from '../components/modals/EmployeeModal';
 import { ViewEmployeeModal } from '../components/modals/ViewEmployeeModal';
 import { ExportModal } from '../components/modals/ExportModal';
-import { Avatar } from '../components/common/Avatar';
+import { EmployeeTable } from '../components/tables/EmployeeTable';
+import { EmployeeFilters } from '../components/filters/EmployeeFilters';
+import { StatusTabs } from '../components/filters/StatusTabs';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Checkbox } from '@/components/ui/checkbox';
 import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Table,
-  TableHeader,
-  TableBody,
-  TableRow,
-  TableHead,
-  TableCell,
   Dialog,
   DialogContent,
   DialogHeader,
@@ -58,17 +21,11 @@ import {
   DialogTitle,
   DialogDescription,
   DialogClose,
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
 } from '@/components/ui';
 import type { PeoplePageProps } from './types';
 import type { EmployeeFilters } from '../types';
 import type { SortOption } from '../types';
 
-const EMPLOYEES_PER_PAGE = 5;
 
 export const PeoplePage: React.FC<PeoplePageProps> = ({
   employees,
@@ -93,8 +50,6 @@ export const PeoplePage: React.FC<PeoplePageProps> = ({
   const [sorting, setSorting] = useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
 
-  const uniqueDepartments = getUniqueDepartments(employees);
-  const uniqueJobTitles = getUniqueJobTitles(employees);
 
   const filters: EmployeeFilters = {
     searchQuery,
@@ -106,184 +61,6 @@ export const PeoplePage: React.FC<PeoplePageProps> = ({
   };
   const filtered = filterEmployeesUtil(employees, filters);
 
-  // Actions Menu Component
-  const ActionsMenu: React.FC<{ employee: Employee }> = ({ employee }) => (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="ghost" size="icon" className="text-gray-400 hover:text-gray-600">
-          <span className="sr-only">Open actions</span>
-          <MoreVertical size={20} />
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-44">
-        <DropdownMenuItem
-          onSelect={e => {
-            e.preventDefault();
-            handleView(employee);
-          }}
-          className="flex items-center gap-2"
-        >
-          <Eye size={16} />
-          View Details
-        </DropdownMenuItem>
-        <DropdownMenuItem
-          onSelect={e => {
-            e.preventDefault();
-            handleEdit(employee);
-          }}
-          className="flex items-center gap-2"
-        >
-          <Edit size={16} />
-          Edit
-        </DropdownMenuItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuItem
-          onSelect={e => {
-            e.preventDefault();
-            handleDelete(employee);
-          }}
-          className="flex items-center gap-2 text-red-600 focus:text-red-700"
-        >
-          <Trash2 size={16} />
-          Delete
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-
-  // Custom global filter function
-  const globalFilterFn = (row: any, _columnId: string, value: string) => {
-    const employee = row.original as Employee;
-    const searchValue = value.toLowerCase();
-
-    // Search in firstName, lastName, and email
-    const firstName = (employee.firstName || '').toLowerCase();
-    const lastName = (employee.lastName || '').toLowerCase();
-    const email = (employee.email || '').toLowerCase();
-    const fullName = `${firstName} ${lastName}`;
-
-    return (
-      firstName.includes(searchValue) ||
-      lastName.includes(searchValue) ||
-      fullName.includes(searchValue) ||
-      email.includes(searchValue)
-    );
-  };
-
-  // Define columns for TanStack Table
-  const columns = useMemo<ColumnDef<Employee>[]>(
-    () => [
-      {
-        id: 'select',
-        header: ({ table }) => (
-          <Checkbox
-            checked={table.getIsAllPageRowsSelected()}
-            onCheckedChange={value => table.toggleAllPageRowsSelected(!!value)}
-            aria-label="Select all"
-          />
-        ),
-        cell: ({ row }) => (
-          <Checkbox
-            checked={row.getIsSelected()}
-            onCheckedChange={value => row.toggleSelected(!!value)}
-            aria-label="Select row"
-          />
-        ),
-        enableSorting: false,
-        enableHiding: false,
-        enableGlobalFilter: false,
-      },
-      {
-        accessorKey: 'name',
-        header: 'Name',
-        cell: ({ row }) => {
-          const employee = row.original;
-          return (
-            <div className="flex items-center gap-3">
-              <Avatar
-                name={`${employee.firstName} ${employee.lastName}`}
-                imageUrl={
-                  employee.profilePhoto || getAvatarUrl(employee.firstName, employee.lastName)
-                }
-                size="sm"
-              />
-              <div>
-                <div className="font-medium text-gray-900">
-                  {employee.firstName} {employee.lastName}
-                </div>
-                <div className="text-sm text-gray-500">{(employee.email || '').toLowerCase()}</div>
-              </div>
-            </div>
-          );
-        },
-        enableSorting: true,
-        sortingFn: (rowA, rowB) => {
-          const nameA = `${rowA.original.firstName} ${rowA.original.lastName}`;
-          const nameB = `${rowB.original.firstName} ${rowB.original.lastName}`;
-          return nameA.localeCompare(nameB);
-        },
-
-        enableGlobalFilter: true,
-        accessorFn: row => `${row.firstName} ${row.lastName} ${row.email}`,
-      },
-      {
-        accessorKey: 'hireDate',
-        header: 'Hire Date',
-        cell: ({ row }) => (
-          <div className="text-sm text-gray-900">{formatDate(row.original.hireDate)}</div>
-        ),
-        enableSorting: true,
-      },
-      {
-        accessorKey: 'jobTitle',
-        header: 'Job Title',
-        cell: ({ row }) => <div className="text-sm text-gray-900">{row.original.jobTitle}</div>,
-        enableSorting: true,
-      },
-      {
-        accessorKey: 'contractType',
-        header: 'Employment Type',
-        cell: ({ row }) => <div className="text-sm text-gray-900">{row.original.contractType}</div>,
-        enableSorting: true,
-        enableGlobalFilter: false,
-      },
-      {
-        id: 'actions',
-        header: '',
-        cell: ({ row }) => <ActionsMenu employee={row.original} />,
-        enableSorting: false,
-        enableGlobalFilter: false,
-      },
-    ],
-    []
-  );
-
-  // Initialize TanStack Table
-  const table = useReactTable({
-    data: filtered,
-    columns,
-    state: {
-      sorting,
-      columnFilters,
-      rowSelection,
-      globalFilter,
-    },
-    enableRowSelection: true,
-    onRowSelectionChange: setRowSelection,
-    onSortingChange: setSorting,
-    onColumnFiltersChange: setColumnFilters,
-    onGlobalFilterChange: setGlobalFilter,
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    globalFilterFn: globalFilterFn,
-    initialState: {
-      pagination: {
-        pageSize: EMPLOYEES_PER_PAGE,
-      },
-    },
-  });
 
   const handleView = (employee: Employee) => {
     setSelectedEmployee(employee);
@@ -351,24 +128,7 @@ export const PeoplePage: React.FC<PeoplePageProps> = ({
       </div>
 
       {/* Status Tabs */}
-      <div className="border-b border-gray-200 overflow-x-auto">
-        <div className="flex gap-8 min-w-max">
-          {statusTabs.map(status => (
-            <Button
-              key={status}
-              variant="ghost"
-              onClick={() => setStatusFilter(status)}
-              className={`py-3 px-1 text-sm font-medium border-b-2 transition-colors whitespace-nowrap rounded-none h-auto ${
-                statusFilter === status
-                  ? 'border-purple-600 text-purple-600'
-                  : 'border-transparent text-gray-500 hover:text-gray-700 hover:bg-transparent'
-              }`}
-            >
-              {status}
-            </Button>
-          ))}
-        </div>
-      </div>
+      <StatusTabs statusFilter={statusFilter} setStatusFilter={setStatusFilter} />
 
       {/* Sick Leave Banner */}
       <div className="bg-gray-50 border border-gray-200 rounded-lg p-4">
@@ -380,220 +140,41 @@ export const PeoplePage: React.FC<PeoplePageProps> = ({
       </div>
 
       {/* Search and Filters */}
-      <div className="space-y-4">
-        <div className="flex flex-col sm:flex-row gap-4">
-          {/* Search Input */}
-          <div className="flex-1 relative">
-            <Search
-              className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
-              size={20}
-            />
-            <Input
-              type="text"
-              placeholder="Search employees..."
-              value={globalFilter ?? ''}
-              onChange={e => setGlobalFilter(e.target.value)}
-              className="pl-10"
-            />
-          </div>
-
-          {/* Filter Button */}
-          <Button
-            variant="outline"
-            onClick={() => setShowFilters(!showFilters)}
-            className="flex items-center gap-2"
-          >
-            <Filter size={16} />
-            Filters
-          </Button>
-        </div>
-
-        {/* Advanced Filters */}
-        {showFilters && (
-          <div className="bg-gray-50 border border-gray-200 rounded-lg p-4">
-            <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
-              <div className="space-y-2">
-                <Label htmlFor="department-filter">Department</Label>
-                <Select
-                  value={departmentFilter || undefined}
-                  onValueChange={value => setDepartmentFilter(value || '')}
-                >
-                  <SelectTrigger id="department-filter">
-                    <SelectValue placeholder="All Departments" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="all">All Departments</SelectItem>
-                    {uniqueDepartments.map(dept => (
-                      <SelectItem key={dept} value={dept}>
-                        {dept}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="employment-type-filter">Employment Type</Label>
-                <Select
-                  value={employmentTypeFilter || undefined}
-                  onValueChange={value =>
-                    setEmploymentTypeFilter(value === 'all' ? '' : value || '')
-                  }
-                >
-                  <SelectTrigger id="employment-type-filter">
-                    <SelectValue placeholder="All Types" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="all">All Types</SelectItem>
-                    <SelectItem value="Permanent">Permanent</SelectItem>
-                    <SelectItem value="Contract">Contract</SelectItem>
-                    <SelectItem value="Intern">Intern</SelectItem>
-                  </SelectContent>
-                </Select>
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="job-title-filter">Job Title</Label>
-                <Select
-                  value={jobTitleFilter || undefined}
-                  onValueChange={value => setJobTitleFilter(value === 'all' ? '' : value || '')}
-                >
-                  <SelectTrigger id="job-title-filter">
-                    <SelectValue placeholder="All Titles" />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="all">All Titles</SelectItem>
-                    {uniqueJobTitles.map(title => (
-                      <SelectItem key={title} value={title}>
-                        {title}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-
-              <div className="flex items-end">
-                <Button
-                  variant="ghost"
-                  onClick={clearFilters}
-                  className="text-gray-600 hover:text-gray-800"
-                >
-                  Clear Filters
-                </Button>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
+      <EmployeeFilters
+        employees={employees}
+        globalFilter={globalFilter}
+        setGlobalFilter={setGlobalFilter}
+        departmentFilter={departmentFilter}
+        setDepartmentFilter={setDepartmentFilter}
+        employmentTypeFilter={employmentTypeFilter}
+        setEmploymentTypeFilter={setEmploymentTypeFilter}
+        jobTitleFilter={jobTitleFilter}
+        setJobTitleFilter={setJobTitleFilter}
+        showFilters={showFilters}
+        setShowFilters={setShowFilters}
+        onClearFilters={clearFilters}
+      />
 
       {/* Employee Table */}
-      <div className="bg-white border border-gray-200 rounded-lg overflow-hidden">
-        <div className="overflow-x-auto">
-          <Table>
-            <TableHeader>
-              {table.getHeaderGroups().map(headerGroup => (
-                <TableRow key={headerGroup.id} className="bg-gray-50">
-                  {headerGroup.headers.map(header => (
-                    <TableHead
-                      key={header.id}
-                      className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-                    >
-                      {header.isPlaceholder ? null : (
-                        <div
-                          className={
-                            header.column.getCanSort()
-                              ? 'cursor-pointer select-none flex items-center gap-2'
-                              : ''
-                          }
-                          onClick={header.column.getToggleSortingHandler()}
-                        >
-                          {flexRender(header.column.columnDef.header, header.getContext())}
-                          {header.column.getCanSort() && (
-                            <ArrowUpDown size={14} className="opacity-50" />
-                          )}
-                        </div>
-                      )}
-                    </TableHead>
-                  ))}
-                </TableRow>
-              ))}
-            </TableHeader>
-            <TableBody>
-              {table.getRowModel().rows.length > 0 ? (
-                table.getRowModel().rows.map(row => (
-                  <TableRow key={row.id} className="hover:bg-gray-50">
-                    {row.getVisibleCells().map(cell => (
-                      <TableCell key={cell.id} className="px-6 py-4">
-                        {flexRender(cell.column.columnDef.cell, cell.getContext())}
-                      </TableCell>
-                    ))}
-                  </TableRow>
-                ))
-              ) : (
-                <TableRow>
-                  <TableCell colSpan={columns.length} className="text-center py-12">
-                    <div className="text-gray-400 mb-2">
-                      <Users size={48} className="mx-auto" />
-                    </div>
-                    <h3 className="text-lg font-medium text-gray-900 mb-1">No employees found</h3>
-                    <p className="text-gray-500">Try adjusting your search or filter criteria.</p>
-                  </TableCell>
-                </TableRow>
-              )}
-            </TableBody>
-          </Table>
-        </div>
-      </div>
-
-      {/* Pagination Controls */}
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-2">
-          <p className="text-sm text-gray-700">
-            Showing {table.getState().pagination.pageIndex * EMPLOYEES_PER_PAGE + 1} to{' '}
-            {Math.min(
-              (table.getState().pagination.pageIndex + 1) * EMPLOYEES_PER_PAGE,
-              table.getFilteredRowModel().rows.length
-            )}{' '}
-            of {table.getFilteredRowModel().rows.length} results
-          </p>
-        </div>
-        <div className="flex items-center gap-2">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.previousPage()}
-            disabled={!table.getCanPreviousPage()}
-          >
-            Previous
-          </Button>
-          <span className="text-sm text-gray-700">
-            Page {table.getState().pagination.pageIndex + 1} of {table.getPageCount()}
-          </span>
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.nextPage()}
-            disabled={!table.getCanNextPage()}
-          >
-            Next
-          </Button>
-        </div>
-      </div>
-
-      {/* Selection Info */}
-      {Object.keys(rowSelection).length > 0 && (
-        <div className="bg-blue-50 border border-blue-200 rounded-lg p-4">
-          <p className="text-sm text-blue-800">
-            {Object.keys(rowSelection).length} employee(s) selected
-          </p>
-        </div>
-      )}
+      <EmployeeTable
+        employees={filtered}
+        globalFilter={globalFilter}
+        sorting={sorting}
+        setSorting={setSorting}
+        columnFilters={columnFilters}
+        setColumnFilters={setColumnFilters}
+        rowSelection={rowSelection}
+        setRowSelection={setRowSelection}
+        onView={handleView}
+        onEdit={handleEdit}
+        onDelete={handleDelete}
+      />
 
       {/* Modals */}
       <ExportModal
         isOpen={showExportModal}
         onClose={() => setShowExportModal(false)}
-        employees={table.getFilteredRowModel().rows.map(row => row.original)}
+        employees={filtered}
       />
 
       <ViewEmployeeModal
