+++ src/utils/index.ts
@@ -1,17 +1,4 @@
-// utils/index.ts
-import type { Employee, EmployeeFilters } from '../types';
-
-export const getAvatarUrl = (firstName: string, lastName: string) =>
-  `https://api.dicebear.com/7.x/initials/png?seed=${encodeURIComponent(`${firstName} ${lastName}`)}`;
-
-export const formatDate = (dateStr: string): string => {
-  const date = new Date(dateStr);
-  return date.toLocaleDateString('en-US', {
-    day: 'numeric',
-    month: 'short',
-    year: 'numeric',
-  });
-};
+import type { Employee, EmployeeFilters } from '@/types';
 
 export const filteredEmployees = (employees: Employee[], filters: EmployeeFilters): Employee[] => {
   let filtered = employees;
@@ -70,54 +57,7 @@ export const filteredEmployees = (employees: Employee[], filters: EmployeeFilter
   return filtered;
 };
 
-export const getUniqueDepartments = (employees: Employee[]): string[] =>
-  [...new Set(employees.map(emp => emp.department))].sort();
-
-export const getUniqueJobTitles = (employees: Employee[]): string[] =>
-  [...new Set(employees.map(emp => emp.jobTitle))].sort();
-
-export const exportToCSV = (data: Employee[]) => {
-  const csvContent = [
-    [
-      'First Name',
-      'Last Name',
-      'Email',
-      'Job Title',
-      'Department',
-      'Employment Type',
-      'Status',
-      'Hire Date',
-    ],
-    ...data.map(emp => [
-      emp.firstName,
-      emp.lastName,
-      emp.email,
-      emp.jobTitle,
-      emp.department,
-      emp.contractType,
-      emp.status,
-      emp.hireDate,
-    ]),
-  ]
-    .map(row => row.join(','))
-    .join('\n');
-
-  const blob = new Blob([csvContent], { type: 'text/csv' });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = 'employees.csv';
-  link.click();
-  URL.revokeObjectURL(url);
-};
-
-export const exportToJSON = (data: Employee[]) => {
-  const jsonContent = JSON.stringify(data, null, 2);
-  const blob = new Blob([jsonContent], { type: 'application/json' });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = 'employees.json';
-  link.click();
-  URL.revokeObjectURL(url);
-};
+// Re-export utilities from their specific modules
+export { getAvatarUrl, formatDate, getUniqueDepartments, getUniqueJobTitles } from './formatting';
+export { exportToCSV, exportToJSON } from './export';
+export { validateFormData, fileToBase64 } from './validation/employee';
